Update project with current local changes

+++ backend/src/main/java/com/new_cafe/app/backend/repository/CategoryRepository.java
@@ -6,11 +6,11 @@ import com.new_cafe.app.backend.entity.Category;
 public interface CategoryRepository {
     List<Category> findAll();
 
-    Category findById(Integer id);
+    Category findById(Long id);
 
     void save(Category category);
 
     void update(Category category);
 
-    void delete(Integer id);
+    void delete(Long id);
 }
